Extract Store::{mark_completed, mark_completed_bulk} into new modules.
The `Store` has gotten too big and difficult to navigate. Reorganising
into `impl` blocks in separate modules to make each unit smaller. The
ack/complete operations have been picked off first as they're nice and
isolated so weren't likely to throw up any complexities.
